# docs/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "OpenAPI definition"
  },
...
  "paths" : {
...
    "/v1/portal/requests" : {
      "post" : {
        "tags" : [ "principal-portal-controller" ],
        "operationId" : "submit",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SubmissionRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "202" : {
            "description" : "Accepted",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SubmissionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/portal/requests/{reference}/verify" : {
      "post" : {
        "tags" : [ "principal-portal-controller" ],
        "operationId" : "verify",
        "parameters" : [ {
          "name" : "reference",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/VerifyRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/VerifiedResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/evaluate" : {
      "post" : {
        "tags" : [ "decision-controller" ],
        "operationId" : "evaluate",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/EvaluateRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EvaluateResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/evaluate/batch" : {
      "post" : {
        "tags" : [ "decision-controller" ],
        "operationId" : "evaluateBatch",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "array",
                "items" : {
                  "$ref" : "#/components/schemas/EvaluateRequest"
                }
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/EvaluateResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/consent" : {
      "post" : {
        "tags" : [ "consent-controller" ],
        "operationId" : "capture",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CaptureRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CaptureResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/receipts/{receiptId}" : {
      "get" : {
        "tags" : [ "receipt-controller" ],
        "operationId" : "receipt",
        "parameters" : [ {
          "name" : "receiptId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ConsentReceipt"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/portal/requests/{reference}" : {
      "get" : {
        "tags" : [ "principal-portal-controller" ],
        "operationId" : "status",
        "parameters" : [ {
          "name" : "reference",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "code",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Status"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/consent/{entityId}/{subjectId}/receipt" : {
      "get" : {
        "tags" : [ "consent-controller" ],
        "operationId" : "receipt_1",
        "parameters" : [ {
          "name" : "entityId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "subjectId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ConsentReceipt"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "SubmissionRequest" : {
        "type" : "object",
        "properties" : {
          "entityId" : {
            "type" : "string",
            "minLength" : 1
          },
          "identifierType" : {
            "type" : "string",
            "enum" : [ "PHONE", "EMAIL", "EMPLOYEE_ID", "CANDIDATE_ID", "DEVICE_ID", "EXTERNAL_ID" ]
          },
          "identifierValue" : {
            "type" : "string",
            "maxLength" : 320,
            "minLength" : 0
          },
          "requestType" : {
            "type" : "string",
            "enum" : [ "ACCESS", "CORRECTION", "COMPLETION", "ERASURE", "NOMINATION", "GRIEVANCE", "CONSENT_WITHDRAWAL", "PORTABILITY", "OPT_OUT_OF_SALE" ]
          },
          "jurisdiction" : {
            "type" : "string",
            "enum" : [ "IN", "EU", "UK", "SG", "MY", "KR", "US_CA", "US_CO", "US_CT", "US_TX", "US_OR", "US_MT", "US_DE", "US_NJ", "US_NE", "US_NH", "US_MN", "US_MD", "US_VA", "US_UT", "US_IA", "OTHER" ]
          }
        },
        "required" : [ "entityId", "identifierType", "identifierValue", "requestType" ]
      },
      "SubmissionResponse" : {
        "type" : "object",
        "properties" : {
          "reference" : {
            "type" : "string"
          },
          "expiresAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "message" : {
            "type" : "string"
          }
        }
      },
      "VerifyRequest" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "string",
            "maxLength" : 32,
            "minLength" : 0
          }
        },
        "required" : [ "code" ]
      },
      "VerifiedResponse" : {
        "type" : "object",
        "properties" : {
          "reference" : {
            "type" : "string"
          },
          "requestId" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "receivedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "dueAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "message" : {
            "type" : "string"
          }
        }
      },
      "EvaluateRequest" : {
        "type" : "object",
        "properties" : {
          "entityId" : {
            "type" : "string",
            "minLength" : 1
          },
          "subjectId" : {
            "type" : "string",
            "minLength" : 1
          },
          "purposeCode" : {
            "type" : "string",
            "minLength" : 1
          },
          "channel" : {
            "type" : "string",
            "enum" : [ "WEB", "MOBILE_APP", "VOICE_CALL", "SMS", "WHATSAPP", "EMAIL", "KIOSK", "PAPER", "API", "BULK_IMPORT", "POSTAL" ]
          },
          "jurisdiction" : {
            "type" : "string",
            "enum" : [ "IN", "EU", "UK", "SG", "MY", "KR", "US_CA", "US_CO", "US_CT", "US_TX", "US_OR", "US_MT", "US_DE", "US_NJ", "US_NE", "US_NH", "US_MN", "US_MD", "US_VA", "US_UT", "US_IA", "OTHER" ]
          },
          "applicationId" : {
            "type" : "string"
          },
          "clientId" : {
            "type" : "string"
          },
          "campaignId" : {
            "type" : "string"
          },
          "vendorId" : {
            "type" : "string"
          },
          "at" : {
            "type" : "string",
            "format" : "date-time"
          },
          "context" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            }
          }
        },
        "required" : [ "entityId", "purposeCode", "subjectId" ]
      },
      "EvaluateResponse" : {
        "type" : "object",
        "properties" : {
          "outcome" : {
            "type" : "string"
          },
          "reason" : {
            "type" : "string"
          },
          "explanation" : {
            "type" : "string"
          },
          "legalBasis" : {
            "type" : "string",
            "enum" : [ "CONSENT", "INFERRED_CONSENT", "LEGITIMATE_USE_EMPLOYMENT", "LEGITIMATE_USE_VOLUNTARY", "CONTRACT_PERFORMANCE", "LEGAL_OBLIGATION", "LEGITIMATE_INTEREST", "VITAL_INTEREST", "PUBLIC_INTEREST" ]
          },
          "purposeCode" : {
            "type" : "string"
          },
          "purposeVersion" : {
            "type" : "integer",
            "format" : "int32"
          },
          "policyVersion" : {
            "type" : "string"
          },
          "evaluatedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "consentExpiresAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "obligations" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }
      },
      "CaptureRequest" : {
        "type" : "object",
        "properties" : {
          "entityId" : {
            "type" : "string",
            "minLength" : 1
          },
          "subjectId" : {
            "type" : "string"
          },
          "subject" : {
            "$ref" : "#/components/schemas/SubjectRef"
          },
          "alsoKnownAs" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/SubjectRef"
            }
          },
          "jurisdiction" : {
            "type" : "string",
            "enum" : [ "IN", "EU", "UK", "SG", "MY", "KR", "US_CA", "US_CO", "US_CT", "US_TX", "US_OR", "US_MT", "US_DE", "US_NJ", "US_NE", "US_NH", "US_MN", "US_MD", "US_VA", "US_UT", "US_IA", "OTHER" ]
          },
          "languageTag" : {
            "type" : "string",
            "minLength" : 1
          },
          "channel" : {
            "type" : "string",
            "enum" : [ "WEB", "MOBILE_APP", "VOICE_CALL", "SMS", "WHATSAPP", "EMAIL", "KIOSK", "PAPER", "API", "BULK_IMPORT", "POSTAL" ]
          },
          "applicationId" : {
            "type" : "string"
          },
          "captureMethod" : {
            "type" : "string",
            "enum" : [ "CHECKBOX_OPT_IN", "CLICK_THROUGH", "DOUBLE_OPT_IN", "OTP_VERIFIED", "VERBAL_RECORDED", "WET_SIGNATURE", "BIOMETRIC_ENROLMENT", "PARENTAL_VERIFIED", "RELAYED_BY_CONSENT_MANAGER", "IMPORTED_WITH_PROVENANCE", "INFERRED_FROM_RELATIONSHIP", "NOT_APPLICABLE" ]
          },
          "actorType" : {
            "type" : "string",
            "enum" : [ "SUBJECT", "PARENT_GUARDIAN", "ADMIN", "AGENT", "SYSTEM", "IMPORT", "REGULATOR", "CONSENT_MANAGER" ]
          },
          "actorId" : {
            "type" : "string"
          },
          "noticeId" : {
            "type" : "string"
          },
          "noticeVersion" : {
            "type" : "integer",
            "format" : "int32"
          },
          "choices" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/PurposeChoiceDto"
            },
            "minItems" : 1
          },
          "rejectAllOffered" : {
            "type" : "boolean"
          },
          "occurredAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "idempotencyKey" : {
            "type" : "string"
          },
          "evidenceRef" : {
            "type" : "string"
          },
          "attributes" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            }
          },
          "guardianVerification" : {
            "$ref" : "#/components/schemas/GuardianVerificationDto"
          }
        },
        "required" : [ "actorType", "captureMethod", "choices", "entityId", "jurisdiction", "languageTag" ]
      },
      "GuardianVerificationDto" : {
        "type" : "object",
        "properties" : {
          "method" : {
            "type" : "string",
            "enum" : [ "EXISTING_VERIFIED_ACCOUNT", "DIGILOCKER_VIRTUAL_TOKEN", "OTHER_DOCUMENTED" ]
          },
          "reference" : {
            "type" : "string",
            "minLength" : 1
          },
          "verifiedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "verifiedBy" : {
            "type" : "string"
          }
        },
        "required" : [ "method", "reference", "verifiedAt" ]
      },
      "PurposeChoiceDto" : {
        "type" : "object",
        "properties" : {
          "purposeCode" : {
            "type" : "string",
            "minLength" : 1
          },
          "granted" : {
            "type" : "boolean"
          },
          "preTicked" : {
            "type" : "boolean"
          },
          "separateAction" : {
            "type" : "boolean"
          }
        },
        "required" : [ "purposeCode" ]
      },
      "SubjectRef" : {
        "type" : "object",
        "properties" : {
          "identifierType" : {
            "type" : "string",
            "enum" : [ "PHONE", "EMAIL", "EMPLOYEE_ID", "CANDIDATE_ID", "DEVICE_ID", "EXTERNAL_ID" ]
          },
          "value" : {
            "type" : "string",
            "minLength" : 1
          }
        },
        "required" : [ "identifierType", "value" ]
      },
      "CaptureResponse" : {
        "type" : "object",
        "properties" : {
          "accepted" : {
            "type" : "boolean"
          },
          "subjectId" : {
            "type" : "string"
          },
          "events" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/EventSummary"
            }
          },
          "violations" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ViolationDto"
            }
          }
        }
      },
      "EventSummary" : {
        "type" : "object",
        "properties" : {
          "eventId" : {
            "type" : "string"
          },
          "purposeCode" : {
            "type" : "string"
          },
          "purposeVersion" : {
            "type" : "integer",
            "format" : "int32"
          },
          "eventType" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string",
            "enum" : [ "GRANTED", "DENIED", "WITHDRAWN", "NOT_ASKED", "EXPIRED", "INVALIDATED", "PENDING_SYNC", "CONFLICTED", "UNKNOWN" ]
          },
          "occurredAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "expiresAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "sequenceNumber" : {
            "type" : "integer",
            "format" : "int64"
          },
          "eventHash" : {
            "type" : "string"
          }
        }
      },
      "ViolationDto" : {
        "type" : "object",
        "properties" : {
          "purposeCode" : {
            "type" : "string"
          },
          "code" : {
            "type" : "string"
          },
          "detail" : {
            "type" : "string"
          }
        }
      },
...
      "ConsentReceipt" : {
        "type" : "object",
        "properties" : {
          "schemaVersion" : {
            "type" : "string"
          },
          "receiptId" : {
            "type" : "string"
          },
          "consentRecordId" : {
            "type" : "string"
          },
          "consentManagerRegistrationId" : {
            "type" : "string"
          },
          "issuedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "fiduciaryName" : {
            "type" : "string"
          },
          "fiduciaryId" : {
            "type" : "string"
          },
          "dpoContact" : {
            "type" : "string"
          },
          "subjectId" : {
            "type" : "string"
          },
          "jurisdiction" : {
            "type" : "string",
            "enum" : [ "IN", "EU", "UK", "SG", "MY", "KR", "US_CA", "US_CO", "US_CT", "US_TX", "US_OR", "US_MT", "US_DE", "US_NJ", "US_NE", "US_NH", "US_MN", "US_MD", "US_VA", "US_UT", "US_IA", "OTHER" ]
          },
          "languageTag" : {
            "type" : "string"
          },
          "noticeId" : {
            "type" : "string"
          },
          "noticeVersion" : {
            "type" : "integer",
            "format" : "int32"
          },
          "entries" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Entry"
            }
          },
          "withdrawalUri" : {
            "type" : "string"
          },
          "rightsUri" : {
            "type" : "string"
          },
          "grievanceUri" : {
            "type" : "string"
          },
          "evidenceHash" : {
            "type" : "string"
          },
          "parentalVerification" : {
            "type" : "string",
            "enum" : [ "EXISTING_VERIFIED_ACCOUNT", "DIGILOCKER_VIRTUAL_TOKEN", "OTHER_DOCUMENTED" ]
          }
        }
      },
      "Entry" : {
        "type" : "object",
        "properties" : {
          "purposeCode" : {
            "type" : "string"
          },
          "purposeVersion" : {
            "type" : "integer",
            "format" : "int32"
          },
          "purposeName" : {
            "type" : "string"
          },
          "dataCategories" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "legalBasis" : {
            "type" : "string",
            "enum" : [ "CONSENT", "INFERRED_CONSENT", "LEGITIMATE_USE_EMPLOYMENT", "LEGITIMATE_USE_VOLUNTARY", "CONTRACT_PERFORMANCE", "LEGAL_OBLIGATION", "LEGITIMATE_INTEREST", "VITAL_INTEREST", "PUBLIC_INTEREST" ]
          },
          "status" : {
            "type" : "string",
            "enum" : [ "GRANTED", "DENIED", "WITHDRAWN", "NOT_ASKED", "EXPIRED", "INVALIDATED", "PENDING_SYNC", "CONFLICTED", "UNKNOWN" ]
          },
          "grantedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "expiresAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "withdrawnAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "recipients" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "crossBorderCountries" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "retentionPeriod" : {
            "type" : "string"
          },
          "sensitive" : {
            "type" : "boolean"
          }
        }
      },
...
      "Status" : {
        "type" : "object",
        "properties" : {
          "reference" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "requestType" : {
            "type" : "string"
          },
          "receivedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "dueAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
    }
  }
}